feat(uart): a naive uart driver for my own SBI

+++ os/src/console.rs
@@ -6,7 +6,7 @@ struct Stdout;
 impl Write for Stdout {
     fn write_str(&mut self, s: &str) -> fmt::Result {
         for c in s.chars() {
-            console_putchar(c as usize);
+            console_putchar(c as u8);
         }
         Ok(())
     }
